+++ .eslintrc.js
@@ -1,7 +1,6 @@
 module.exports = {
   extends: ['molindo/javascript'],
   env: {
-    browser: true,
     es6: true,
     jest: true,
     node: true